requirements.txt
amqp==2.5.2
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.4
django-admin-rangefilter==0.5.4
django-appconf==1.0.4
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-on-heroku==1.0.1
django-registration==2.4.1
django-responsive==0.3.0
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.11.0
idna==2.9
kamene==0.32
kombu==4.6.8
mpesa-api==0.1.9
Pillow==8.2.0
psycopg2-binary==2.8.5
python-decouple==3.4
pytz==2021.1
requests==2.23.0
scapy==2.4.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.25.8
vine==1.3.0
whitenoise==5.2.0
